app/requirements.txt
Flask==0.12.1
Jinja2==2.9.6
MySQL-python==1.2.5
python-memcached==1.58
redis==2.10.5

